イテレータモジュールを整頓した
